Refactoring related to clang-tidy

+++ src/tucucore/dosage.cpp
@@ -223,7 +223,7 @@ std::vector<Duration> SimpleDoseList::getTimeStepList(DateTime const& _intervalS
             }
             avgVal += (next->getDateTime() - current->getDateTime()).toMilliseconds();
         }
-        avgVal /= (m_dosageList.size() - 1);
+        avgVal /= static_cast<int64_t>(m_dosageList.size() - 1);
         if (m_dosageList.back()->getDateTime() >= _intervalStart) {
             timeStepList.emplace_back(Duration(std::chrono::milliseconds(avgVal)));
         }
